refactor: Update routing
