+++ server.js
@@ -223,7 +223,7 @@ app.prepare().then(() => {
      */
     server.get('/maker', (req, res) => renderAndCache(req, res, '/maker'))
     server.get('/maker/apps', (req, res) => renderAndCache(req, res, '/maker/apps'))
-    server.get('/maker/apps/:appId', (req, res) => renderAndCache(req, res, '/maker/apps'))
+    server.get('/maker/apps/:appId', (req, res) => renderAndCache(req, res, '/maker/apps/:appId'))
     server.get('/maker/:accessToken', (req, res) => renderAndCache(req, res, '/maker/magic-link', { accessToken: req.params.accessToken }))
 
     apps.platforms.forEach((platform) => {
